Update tf.keras to keras 2.1.5 version.
PiperOrigin-RevId: 191914904
